# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTS OCR API",
    "description": "Cheque Truncation System OCR (Optical Character Recognition) - Mobile and Web API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "CTS Team",
      "email": "support@cts.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5002",
      "description": "Development Server"
    },
    {
      "url": "http://182.76.112.78:91",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      }
    },
    "schemas": {
...
      "RejectionStatusRequest": {
        "type": "object",
        "required": ["presentingMICR", "scannerID", "presentaionDate"],
        "properties": {
          "presentingMICR": {
            "type": "string",
            "example": "411176003"
          },
          "scannerID": {
            "type": "string",
            "example": "70002"
          },
          "presentaionDate": {
            "type": "string",
            "format": "date",
            "example": "2026-05-23"
          }
        }
      },
      "NPCIStatusRequest": {
        "type": "object",
        "required": ["date"],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-23"
          },
          "scannerId": {
            "type": "string",
            "example": "70002"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/cts/getRejectionReport": {
      "post": {
        "tags": ["Reporting"],
        "summary": "Get Rejection Report",
        "description": "Generate report of rejected cheques",
        "operationId": "getRejectionReport",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rejection report generated successfully"
          }
        },
        "x-status": "ACTIVE"
      }
    },
    "/cts/getRetRejectionReport": {
      "post": {
        "tags": ["Reporting"],
        "summary": "Get Rejection Report (retention DB)",
        "description": "Retrieve rejection report directly from the current database instead of the external API",
        "operationId": "getRetRejectionReport",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rejection report generated successfully"
          }
        },
        "x-status": "ACTIVE"
      }
    },
    "/cts/getNPCIStatusReport": {
      "post": {
        "tags": ["Reporting"],
        "summary": "Get NPCI Status Report",
        "description": "Generate NPCI (National Payments Corporation of India) status report",
        "operationId": "getNPCIStatusReport",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NPCIStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NPCI status report generated successfully"
          }
        },
        "x-status": "ACTIVE"
      }
    },
    "/cts/getRejectionReport_v2": {
      "post": {
        "tags": ["Reporting"],
        "summary": "Get Rejection Report (v2)",
        "description": "Generate rejection report - Enhanced version",
        "operationId": "getRejData",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rejection report v2 generated successfully"
          }
        },
        "x-status": "ACTIVE"
      }
    },
...
  }
}